Space Invaders 2D
This game is built in JavaScript. The aliens, moving mothership, the shields and the player spaceship is created using bitmap array which is basically an array of 0's and 1's. The places that are marked with 1's are displayed using colors. The bullets being shot by the player spaceship, aliens and mothership are designed using  object pooling design pattern.